New Way of Checking What Objects to Click
There is now an array on GameManager of tags that are clickable - we have to manually add them to that for now. When a player raycasts to an object, it checks if that gameObject's tag exists in the array, and only then lets you click and does all the other stuff. MUCH better than adding many parameters.